refactor: 收口平台与诊断基础设施
