+++ xl_updata_tool/app/platform/database_compat.py
@@ -1,0 +1,7 @@
+"""数据库平台契约。
+
+当前实现保留在 ``app.core.database``，这里提供稳定的迁移入口。
+"""
+
+from app.core.database import *  # noqa: F401,F403
+
