+++ xl_updata_tool/app/platform/files_compat.py
@@ -1,0 +1,6 @@
+"""文件发布与原子替换的平台契约。"""
+
+from app.core.file_utils import atomic_write_bytes, replace_directory
+
+__all__ = ["atomic_write_bytes", "replace_directory"]
+
