+++ xl_updata_tool/app/platform/processes_compat.py
@@ -1,0 +1,6 @@
+"""外部进程调用的平台契约。"""
+
+from app.core.process_runner import run_external_process
+
+__all__ = ["run_external_process"]
+
